--- Assets/Scripts/Managers/GameManager1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 973c349da5d65c14a92771a8562237da
