--- Assets/Scripts/Managers/MapManager1.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 352319309d6afc044b0ce94699fbbde6
